# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synnovatify User API",
    "description": "Collects user details and stores them in Supabase.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://api.synnovatify.com", "description": "Production" },
    { "url": "http://127.0.0.1:8000", "description": "Local development" }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["health"],
        "summary": "Root",
        "description": "Simple liveness/info endpoint.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "type": "object" },
                "example": { "service": "Synnovatify User API", "status": "ok" }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["health"],
        "summary": "Health",
        "description": "Health check used by Render.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "type": "object" },
                "example": { "status": "healthy" }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": ["users"],
        "summary": "Create User",
        "description": "Validate incoming user details and persist them to Supabase.",
        "operationId": "create_user_api_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRead" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          },
          "502": {
            "description": "Database error while saving the user"
          }
        }
      },
      "get": {
        "tags": ["users"],
        "summary": "List Users",
        "description": "Return the most recent users (newest first).",
        "operationId": "list_users_api_users_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "description": "Clamped server-side to 1-200."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserRead" }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          },
          "502": {
            "description": "Database error while reading users"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreate": {
        "type": "object",
        "title": "UserCreate",
        "description": "Payload accepted when creating a user.",
        "required": ["name", "email"],
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "minLength": 1,
            "maxLength": 200,
            "examples": ["Ada Lovelace"]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": ["ada@example.com"]
          },
          "phone": {
            "anyOf": [
              { "type": "string", "maxLength": 32 },
              { "type": "null" }
            ],
            "title": "Phone",
            "default": null,
            "examples": ["+1-555-0100"]
          },
          "message": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Message",
            "default": null,
            "examples": ["I'd like to know more."]
          }
        }
      },
      "UserRead": {
        "type": "object",
        "title": "UserRead",
        "description": "Representation of a stored user returned by the API.",
        "required": ["id", "name", "email"],
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "phone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Phone",
            "default": null
          },
          "message": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Message",
            "default": null
          },
          "created_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Created At",
            "default": null
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "anyOf": [{ "type": "string" }, { "type": "integer" }]
            },
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        }
      },
      "HTTPValidationError": {
        "type": "object",
        "title": "HTTPValidationError",
        "properties": {
          "detail": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "title": "Detail"
          }
        }
      }
    }
  }
}